邹睿明：订单模块

+++ guns/guns-api/src/main/java/com/stylefeng/guns/order/common/persistence/model/MoocOrder.java
@@ -28,7 +28,7 @@ public class MoocOrder implements Serializable {
 
     private Double orderPrice;
 
-    private Integer orderTimestamp;
+    private Long orderTimestamp;
 
     private String orderStatus;
 }
